chore: move helpers to composable
